+++ examples/mvn-hello-world/pom.xml
@@ -10,11 +10,11 @@
   <url>http://cloudflow.io</url>
 
   <properties>
-    <docker.maven.plugin.version>0.35.0</docker.maven.plugin.version>
+    <docker.maven.plugin.version>0.35-SNAPSHOT</docker.maven.plugin.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <scala.version>2.12.12</scala.version>
-    <cloudflow.version>2.0.26-RC42</cloudflow.version>
+    <cloudflow.version>${env.CLOUDFLOW_VERSION}</cloudflow.version>
     <cloudflow.plugin.version>2.1.0-SNAPSHOT</cloudflow.plugin.version>
   </properties>
